# contracts/openapi/pie-control-plane-v1.yaml
openapi: 3.1.2
info:
  title: Pie Control Plane API
  version: 0.1.0
  description: Initial executable contract for instance discovery, session, organizations, projects, and work items.
servers:
  - url: https://api.pielab.ai
security:
  - PieOidc: []
...
paths:
...
  /v1/session:
    get:
      tags: [Session]
      operationId: getSessionState
      responses:
        '200':
          description: Current Pie account, organization, and permission state without raw tokens.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/session-state.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/organizations/{organizationId}/memberships:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [Memberships]
      operationId: listMemberships
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Memberships visible to the caller.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/resources/membership.v1.schema.json
                  nextCursor:
                    type: [string, 'null']
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/organizations/{organizationId}/meetings/{meetingId}/decisions:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: meetingId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Meetings]
      operationId: listMeetingDecisions
      responses:
        '200':
          description: Reviewable meeting decisions.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/resources/meeting-decision.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/meeting-decisions/{decisionId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: decisionId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    patch:
      tags: [Meetings]
      operationId: updateMeetingDecision
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/resources/meeting-decision-update.v1.schema.json
      responses:
        '200':
          description: Updated decision, reset to unreviewed.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-decision.v1.schema.json
        '409':
          $ref: '#/components/responses/Problem'
        '428':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/meeting-decisions/{decisionId}:review:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: decisionId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    post:
      tags: [Meetings]
      operationId: reviewMeetingDecision
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/resources/meeting-outcome-review.v1.schema.json
      responses:
        '200':
          description: Reviewed decision.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-decision.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    PieOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.pielab.ai/realms/pie/.well-known/openid-configuration
      description: The actual issuer is discovered from /.well-known/pie for each instance.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        $ref: ../schemas/common/opaque-id.v1.schema.json
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
  headers:
    ETag:
      description: Strong entity tag for the returned resource version.
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
...
  responses:
    Problem:
      description: RFC 9457 problem with a stable Pie error code.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/common/problem-details.v1.schema.json